optimize file transfer buffers (#19)

+++ include/network/socketUtils.h
@@ -4,7 +4,7 @@ int createServerSocket(const int port);
 int acceptClient(int serverSocketFd);
 int connectToServer(const char* ip, const int port);
 void closeSocket(int socketFd);
-ssize_t sendAll(int socketFd, const char* buffer, int byteRead);
+ssize_t sendAll(int socketFd, const char* buffer, size_t byteRead);
 
 bool setSocketRecvTimeout(int socketFd, int timeoutSeconds);
 bool setSocketSendTimeout(int socketFd, int timeoutSeconds);
